Separate double modules: FormstackMedliabModule + MedicalBypassModule
