+++ esp/esp/program/modules/handlers/formstackmedliabmodule.py
@@ -32,20 +32,17 @@ Learning Unlimited, Inc.
   Phone: 617-379-0178
   Email: web-team@learningu.org
 """
-from esp.program.modules.base import ProgramModuleObj, needs_teacher, needs_student, needs_admin, usercheck_usetl, meets_deadline, main_call, aux_call, meets_cap
-from esp.program.modules import module_ext
-from esp.utils.web       import render_to_response
-from esp.middleware      import ESPError
-from esp.users.models    import ESPUser, Permission, Record, User
+from esp.program.modules.base import ProgramModuleObj, needs_student, meets_deadline, main_call, aux_call, meets_cap
+from esp.utils.web import render_to_response
+from esp.users.models import ESPUser, Record
 from esp.tagdict.models import Tag
-from django.db.models.query       import Q
-from django.shortcuts import redirect
+from django.db.models.query import Q
 from esp.middleware.threadlocalrequest import get_current_request
-from esp.users.forms.generic_search_form import GenericSearchForm
+
 
 # hackish solution for Splash 2012
 class FormstackMedliabModule(ProgramModuleObj):
-    """ Module for collecting medical information online via Formstack """
+    """Module for collecting medical information online via Formstack."""
 
     @classmethod
     def module_properties(cls):
@@ -55,11 +52,6 @@ class FormstackMedliabModule(ProgramModuleObj):
                 "module_type": "learn",
                 "seq": 3,
                 "required": True
-                },
-                {
-                "admin_title": "Formstack Med-liab Bypass Page",
-                "link_title": "Grant Medliab Bypass",
-                "module_type": "manage",
                 }]
 
     def isCompleted(self):
@@ -72,12 +64,11 @@ class FormstackMedliabModule(ProgramModuleObj):
                 event="med_bypass",
                 program=self.program)
 
-    def students(self, QObject = False):
+    def students(self, QObject=False):
         Q_students = Q(record__event="med",
                        record__program=self.program)
         Q_bypass = Q(record__event="med_bypass",
-                       record__program=self.program)
-
+                     record__program=self.program)
 
         if QObject:
             students = Q_students
@@ -102,9 +93,7 @@ class FormstackMedliabModule(ProgramModuleObj):
     @meets_deadline('/FormstackMedliab')
     @meets_cap
     def medliab(self, request, tl, one, two, module, extra, prog):
-        """
-        Landing page redirecting to med-liab form on Formstack.
-        """
+        """Landing page redirecting to med-liab form on Formstack."""
         t = Tag.getProgramTag("formstack_id", self.program)
         v = Tag.getProgramTag("formstack_viewkey", self.program)
         context = {"formstack_id": t, "formstack_viewkey": v}
@@ -114,41 +103,9 @@ class FormstackMedliabModule(ProgramModuleObj):
     @aux_call
     @needs_student
     def medicalpostback581309742(self, request, tl, one, two, module, extra, prog):
-        """
-        Marks student off as completed.
-        """
+        """Marks student off as completed."""
         Record.objects.create(user=request.user, event="med", program=self.program)
         return self.goToCore(tl)
-
-    @main_call
-    @needs_admin
-    def medicalbypass(self, request, tl, one, two, module, extra, prog):
-        status = None
-
-        if request.method == 'POST':
-            form = GenericSearchForm(request.POST)
-            if form.is_valid():
-                user = form.cleaned_data['target_user']
-                if Record.objects.filter(user=user,
-                                          program=self.program,
-                                          event="med_bypass").exists():
-                    status = 'bypass exists'
-                elif Record.objects.filter(user=user,
-                                            program=self.program,
-                                            event="med").exists():
-                    status = 'reg bit exists'
-                else:
-                    Record.objects.create(user=user,
-                                           program=self.program,
-                                           event="med_bypass")
-                    status = 'success'
-            else:
-                status = 'invalid user'
-
-        context = {'status': status, 'form': GenericSearchForm()}
-
-        return render_to_response(self.baseDir()+'medicalbypass.html',
-                                  request, context)
 
     class Meta:
         proxy = True
